+++ Rdmp.Core/DataLoad/Modules/Attachers/RemoteTableAttacher.cs
@@ -37,7 +37,10 @@ public class RemoteTableAttacher : RemoteAttacher
 {
     private const string FutureLoadMessage = "Cannot load data from the future";
 
-    public RemoteTableAttacher() : base()
+    public RemoteTableAttacher(bool requestsExternalDatabaseCreation=true) : base(requestsExternalDatabaseCreation)
+    {
+    }
+    public RemoteTableAttacher() : base(true)
     {
     }