+++ Rdmp.UI/Wizard/SimpleFilterUI.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
@@ -38,6 +39,7 @@ public partial class SimpleFilterUI : UserControl
 
     public IFilter Filter => _filter;
 
+    [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
     public bool Mandatory
     {
         get => _mandatory;