+++ CHANGELOG.md
@@ -4,9 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [8.4.1] - Unreleased
+## [Unreleased]
 
+- Build on and target .Net 9 rather than 8
 - Simplify DB Patching Interface
+
+## [8.4.2] - 2024-12-18
+
+- Fix issue with MEF constructing Remote Table Attachers
+
+## [8.4.1] - 2024-12-10
+
+- Improve Overview Model Generation Speed
+- Add Remote Table Without DB Creation Attacher
+>>>>>>> 188afdb9501892f22bd66d56cb8d095c49a14fb0
 
 ## [8.4.0] - 2024-12-02